+++ src/RestApi/SettingsRestController.php
@@ -19,6 +19,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
+use Symfony\Component\Routing\Requirement\Requirement;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class SettingsRestController extends AbstractFoodsharingRestController
@@ -80,19 +81,18 @@ class SettingsRestController extends AbstractFoodsharingRestController
         return $this->handleView($this->view([], 204));
     }
 
-    /**
-     * Requests that the user's login email address will be changed. This does not permanently change the address yet,
-     * but sends out the confirmation email.
-     */
+    #[OA2\Patch(summary: 'Requests that the user\'s login email address will be changed. This does not permanently
+    change the address yet, but sends out the confirmation email. Every user can change the own email address. Changing
+    someone else\'s address requires certain permissions.')]
     #[OA2\Tag(name: 'user')]
-    #[Rest\Patch('user/current/email')]
+    #[Rest\Patch('user/{userId}/email', requirements: ['userId' => Requirement::POSITIVE_INT])]
     #[ParamConverter('request', converter: 'fos_rest.request_body')]
     #[OA2\RequestBody(content: new Model(type: EmailChangeRequest::class))]
     #[OA2\Response(response: '200', description: 'Success')]
     #[OA2\Response(response: '400', description: 'Empty or invalid parameters')]
     #[OA2\Response(response: '401', description: 'Not logged in')]
     #[OA2\Response(response: '403', description: 'Wrong password')]
-    public function requestEmailChange(EmailChangeRequest $request, ValidatorInterface $validator): Response
+    public function requestEmailChange(int $userId, EmailChangeRequest $request, ValidatorInterface $validator): Response
     {
         if (!$this->session->mayRole()) {
             throw new UnauthorizedHttpException('');
@@ -104,7 +104,11 @@ class SettingsRestController extends AbstractFoodsharingRestController
             throw new BadRequestHttpException(json_encode(['field' => $firstError->getPropertyPath(), 'message' => $firstError->getMessage()]));
         }
 
-        $this->settingsTransactions->requestEmailChange($request);
+        if ($userId == $this->session->id()) {
+            $this->settingsTransactions->requestEmailChange($request);
+        } else {
+            $this->settingsTransactions->changeLoginEmail($request, $userId);
+        }
 
         return $this->handleView($this->view([], 200));
     }